+++ CHANGELOG.md
@@ -4,6 +4,10 @@ All notable changes to thsi project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/).
 
 ## Unversioned
+## Added
+- Due day abbriviations supported
+- Added archive command
+
 ## Changed
 - List command now case insensitive